refactor: extract shared config core with per-analysis profiles
