+++ constants.py
@@ -2,6 +2,8 @@
 Constant definitions.
 """
 
+import enum
+
 __all__ = [
     "GLOBAL_SCOPES",
     "ET_SCOPES",
@@ -18,6 +20,9 @@ __all__ = [
     "HAD_TAU_SCOPES",
     "SCOPES",
     "ERAS",
+    "AvailableBJetIDs",
+    "BJET_ID_ALGORITHM",
+    "LEGACY_AVAILABLE_SAMPLES",
 ]
 
 
@@ -324,4 +329,48 @@ XYH_MASS_POINTS = [
     (4000, 2600),
     (4000, 3000),
     (4000, 3500),
+]
+
+# b jet identification algorithms
+class AvailableBJetIDs(enum.Enum):
+    DEEPJET = enum.auto()
+    PNET = enum.auto()
+    UPART = enum.auto()
+BJET_ID_ALGORITHM = AvailableBJetIDs.PNET
+
+# legacy sample surface (verbatim, formerly generate.py:11-43); the fallback
+# for any config module that defines none of AVAILABLE_SAMPLES/DEFAULT_SAMPLES
+LEGACY_AVAILABLE_SAMPLES = [
+    "ggh_htautau",
+    "ggh_hbb",
+    "vbf_htautau",
+    "vbf_hbb",
+    "rem_htautau",
+    "rem_hbb",
+    "rem_hww",
+    "rem_hzz",
+    "rem_higgs",
+    "higgs",
+    "hh4b",
+    "hh2b2tau",
+    "hh4v",
+    "embedding",
+    "embedding_mc",
+    "singletop",
+    "ttbar",
+    "rem_ttbar",
+    "diboson",
+    "dyjets",
+    "dyjets_madgraph",
+    "dyjets_amcatnlo",
+    "dyjets_amcatnlo_ll",
+    "dyjets_amcatnlo_tt",
+    "dyjets_powheg",
+    "wjets",
+    "wjets_madgraph",
+    "wjets_amcatnlo",
+    "data",
+    "electroweak_boson",
+    "nmssm_Ybb",
+    "nmssm_Ytautau",
 ]
