+++ generate.py
@@ -2,48 +2,29 @@ from os import path
 import importlib
 from code_generation.code_generation import CodeGenerator
 
-from .constants import ERAS, SCOPES
+from .constants import ERAS, SCOPES, LEGACY_AVAILABLE_SAMPLES
+
+
+def resolve_sample_surface(config_module):
+    """Resolve (available_samples, default_samples) for a config module.
+
+    Config modules may optionally define AVAILABLE_SAMPLES and/or
+    DEFAULT_SAMPLES. If neither is defined (e.g. nmssm_config), both fall
+    back to the legacy hardcoded sample list for backward compatibility.
+    """
+    available = getattr(config_module, "AVAILABLE_SAMPLES", None)
+    default = getattr(config_module, "DEFAULT_SAMPLES", None)
+    if available is None and default is None:
+        return list(LEGACY_AVAILABLE_SAMPLES), list(LEGACY_AVAILABLE_SAMPLES)
+    if available is None:
+        available = list(default)
+    if default is None:
+        default = list(available)
+    return list(available), list(default)
 
 
 def run(args):
     analysis_name = "bbtautau"
-
-    available_samples = [
-        "ggh_htautau",
-        "ggh_hbb",
-        "vbf_htautau",
-        "vbf_hbb",
-        "rem_htautau",
-        "rem_hbb",
-        "rem_hww",
-        "rem_hzz",
-        "rem_higgs",
-        "higgs",
-        "hh4b",
-        "hh2b2tau",
-        "hh4v",
-        "embedding",
-        "embedding_mc",
-        "singletop",
-        "ttbar",
-        "rem_ttbar",
-        "diboson",
-        "dyjets",
-        "dyjets_madgraph",
-        "dyjets_amcatnlo",
-        "dyjets_amcatnlo_ll",
-        "dyjets_amcatnlo_tt",
-        "dyjets_powheg",
-        "wjets",
-        "wjets_madgraph",
-        "wjets_amcatnlo",
-        "data",
-        "electroweak_boson",
-        "nmssm_Ybb",
-        "nmssm_Ytautau",
-    ]
-    available_eras = ERAS
-    available_scopes = SCOPES
 
     ## setup variables
     shifts = set([shift.lower() for shift in args.shifts])
@@ -56,6 +37,23 @@ def run(args):
     config = importlib.import_module(
         f"analysis_configurations.{analysis_name}.{configname}"
     )
+
+    ## resolve and enforce the config-specific sample/era surface BEFORE
+    ## build_config is invoked
+    available_samples, _ = resolve_sample_surface(config)
+    available_eras = getattr(config, "AVAILABLE_ERAS", ERAS)
+    available_scopes = SCOPES
+    if era not in available_eras:
+        raise ValueError(
+            f"Config '{configname}' does not support era '{era}' "
+            f"(supported: {available_eras})."
+        )
+    if sample_group not in available_samples:
+        raise ValueError(
+            f"Config '{configname}' does not accept sample '{sample_group}' "
+            f"(accepted: {available_samples})."
+        )
+
     ## Setting up executable
     executable = f"{configname}_{sample_group}_{era}.cxx"
     args.logger.info(f"Generating code for {sample_group}...")
